requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
ascii-magic==1.6
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
discord.py==2.0.1
dnspython==2.2.1
Flask==2.2.2
frozenlist==1.3.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
Pillow==9.2.0
psutil==5.9.2
pycparser==2.21
pymongo==4.2.0
PyNaCl==1.5.0
python-dotenv==0.21.0
termcolor==2.0.1
Werkzeug==2.2.2
yarl==1.8.1
